requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.11
colorama==0.4.4
cryptocompare==0.7.5
filelock==3.4.2
idna==3.3
Pillow==9.0.1
psutil==5.9.0
python-monerorpc==0.6.2
qrcode==7.3.1
requests==2.27.1
urllib3==1.26.8
